Revert the shared timeseries provider, fix the hook in place
The provider fetched every key through one `trpc.useQueries` call built from a
freshly allocated array each render. React Query memoises `defaultedQueries` on
that array's identity (useQueries.js:44), so `setQueries` ran every render, the
observer republished, and `useSyncExternalStore` saw a new snapshot each pass -
the same notifyManager -> forceStoreRerender loop it was meant to fix, just
concentrated in one observer. Confirmed still crashing on preview.

Keep only the root-cause fix: drop the hand-rolled Map, the forced re-render
counter, and the `Date.now()`-derived query options, so the query key and
options are fixed at mount. The window is bucketed to the minute so both cells
in a row share one cache entry.
