+++ web/apps/dashboard/app/(app)/[workspaceSlug]/apis/[apiId]/keys/[keyAuthId]/_components/components/table/keys-list.tsx
@@ -6,7 +6,6 @@ import {
   useApiKeysListQuery,
 } from "@/components/api-keys-table";
 import { EmptyApiKeys } from "@/components/api-keys-table/components/empty-api-keys";
-import { KeyTimeseriesProvider } from "@/components/api-keys-table/components/key-timeseries-provider";
 import { SelectionControls } from "@/components/api-keys-table/components/selection-controls";
 import { useWorkspaceNavigation } from "@/hooks/use-workspace-navigation";
 import type { KeyDetails } from "@/lib/trpc/routers/api/keys/query-api-keys/schema";
@@ -150,26 +149,22 @@ export const KeysList = ({
 
   const isNavigating = isFetching && !isInitialLoading;
 
-  const keyIds = useMemo(() => keys.map((key) => key.id), [keys]);
-
   return (
     <>
-      <KeyTimeseriesProvider keyAuthId={keyspaceId} keyIds={keyIds}>
-        <DataTable
-          data={keys}
-          columns={columns}
-          getRowId={(key) => key.id}
-          isLoading={isInitialLoading}
-          onRowClick={handleRowClick}
-          selectedItem={selectedKey}
-          rowClassName={getRowClassNameMemoized}
-          renderSkeletonRow={renderApiKeySkeletonRow}
-          sorting={sorting}
-          onSortingChange={handleSortingChange}
-          config={TABLE_CONFIG}
-          emptyState={<EmptyApiKeys />}
-        />
-      </KeyTimeseriesProvider>
+      <DataTable
+        data={keys}
+        columns={columns}
+        getRowId={(key) => key.id}
+        isLoading={isInitialLoading}
+        onRowClick={handleRowClick}
+        selectedItem={selectedKey}
+        rowClassName={getRowClassNameMemoized}
+        renderSkeletonRow={renderApiKeySkeletonRow}
+        sorting={sorting}
+        onSortingChange={handleSortingChange}
+        config={TABLE_CONFIG}
+        emptyState={<EmptyApiKeys />}
+      />
       <PaginationFooter
         page={page}
         pageSize={pageSize}
