+++ web/apps/dashboard/components/api-keys-table/columns/create-api-key-columns.tsx
@@ -262,7 +262,13 @@ export const createApiKeyColumns = ({
     },
     cell: ({ row }) => {
       const key = row.original;
-      return <VerificationBarChart keyId={key.id} selected={key.id === selectedKeyId} />;
+      return (
+        <VerificationBarChart
+          keyAuthId={keyspaceId}
+          keyId={key.id}
+          selected={key.id === selectedKeyId}
+        />
+      );
     },
   },
   {
@@ -317,7 +323,9 @@ export const createApiKeyColumns = ({
     },
     cell: ({ row }) => {
       const key = row.original;
-      return <StatusDisplay keyData={key} isSelected={key.id === selectedKeyId} />;
+      return (
+        <StatusDisplay keyData={key} keyAuthId={keyspaceId} isSelected={key.id === selectedKeyId} />
+      );
     },
   },
   {
