+++ web/apps/dashboard/components/api-keys-table/components/status-cell/index.tsx
@@ -7,11 +7,12 @@ import { useKeyStatus } from "./use-key-status";
 
 type StatusDisplayProps = {
   keyData: KeyDetails;
+  keyAuthId: string;
   isSelected: boolean;
 };
 
-export const StatusDisplay = ({ keyData, isSelected }: StatusDisplayProps) => {
-  const { primary, count, isLoading, statuses, isError } = useKeyStatus(keyData);
+export const StatusDisplay = ({ keyAuthId, keyData, isSelected }: StatusDisplayProps) => {
+  const { primary, count, isLoading, statuses, isError } = useKeyStatus(keyAuthId, keyData);
   const enableKeyMutation = useUpdateKeyStatus();
 
   if (isLoading) {
